STYLE: Refactor support of itkFloatingPointExceptions implementations
